update generate and readme
