+++ llava/utils.py
@@ -1,5 +1,9 @@
 from pathlib import Path
 from huggingface_hub import snapshot_download
+import os
+import requests
+from PIL import Image
+import mlx.core as mx
 
 
 def get_model_path(path_or_hf_repo: str) -> Path:
@@ -28,3 +32,30 @@ def get_model_path(path_or_hf_repo: str) -> Path:
             )
         )
     return model_path
+
+
+def load_image(image_source):
+    if image_source.startswith(("http://", "https://")):
+        try:
+            response = requests.get(image_source, stream=True)
+            response.raise_for_status()
+            return Image.open(response.raw)
+        except requests.HTTPError as e:
+            print(f"Failed to load image from URL: {e}")
+            return None
+    elif os.path.isfile(image_source):
+        try:
+            return Image.open(image_source)
+        except IOError as e:
+            print(f"Failed to load image from path: {e}")
+            return None
+    else:
+        print("The image source is neither a valid URL nor a file path.")
+        return None
+
+
+def prepare_inputs(processor, image, prompt):
+    inputs = processor(prompt, image, return_tensors="np")
+    pixel_values = mx.array(inputs["pixel_values"])
+    input_ids = mx.array(inputs["input_ids"])
+    return input_ids, pixel_values
